requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
scipy==1.11.4
dill
mlflow
pyyaml==6.0.1
